source/gateway/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2==2.9.2
requests==2.26.0
